factor out common poa code and exceptions
