+++ web3_multi_provider/exceptions.py
@@ -1,0 +1,8 @@
+class NoActiveProviderError(Exception):
+    """Base exception if all providers are offline"""
+
+
+class ProtocolNotSupported(Exception):
+    """Supported protocols: http, https"""
+
+
